#542 Moves login endpoint to UserController
